% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reads.R
\name{biotek_300_700_universal}
\alias{biotek_300_700_universal}
\title{helper for the 300:700 nm spectra biotek read protocol}
\usage{
biotek_300_700_universal(file)
}
\arguments{
\item{file}{Path to the input file}
}
\description{
This function parses a read file from a spectrophotometer
}
