% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reads.R
\name{biotek_ten_wavelengths_universal}
\alias{biotek_ten_wavelengths_universal}
\title{helper for the ten_wavelengths.prt protocol}
\usage{
biotek_ten_wavelengths_universal(file)
}
\arguments{
\item{file}{Path to the input file}
}
\description{
This function parses a read file from a spectrophotometer
}
