% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{FO_TWI_lin_model}
\alias{FO_TWI_lin_model}
\title{Train a linear model with First order and Two way interaction terms}
\usage{
FO_TWI_lin_model(abs, pH)

FO_TWI_lin_model(abs, pH)
}
\arguments{
\item{abs}{a n x k data frame of absorbance measurements, where n is the number of observations, and k is the number of wavelengths}

\item{ph}{a normalized vector of length n of pH measurements for training}
}
\value{
a generated function that makes a prediction with the model with new data as an input

a generated function that makes a prediction with the model with new data as an input
}
\description{
pH ~ FO(abs) + TWI(abs)

pH ~ FO(abs) + TWI(abs)
}
\details{
This function does not include hyperparameter optimization and assumes that all absorbance inputs and pH measurements are normalized

This function does not include hyperparameter optimization and assumes that all absorbance inputs and pH measurements are normalized
}
