% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_to_feature.R
\name{gff3_to_feature}
\alias{gff3_to_feature}
\title{Read gff3 files into Grange objects}
\usage{
gff3_to_feature(file, max_pval = 1, max_qval = 1)
}
\arguments{
\item{file}{Character. Full path to the gff3 files}

\item{max_pval}{Numeric.}

\item{max_qval}{Numeric.}
}
\value{
Data frame.
}
\description{
Read gff3 files into Grange objects
}
