% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gp_model}
\alias{gp_model}
\title{Train a GP model for making pH predictions}
\usage{
gp_model(abs, ph)
}
\arguments{
\item{abs}{a n x k dataframe of absorbance measurements, where n is the number of observations, and k is the number of wavelengths}

\item{ph}{a normalized vector of length n of pH measurements for training}
}
\value{
a generated function that makes a prediction with the model with new data as an input
}
\description{
This function does not include hyperparameter optimization and assumes that all absorbance inputs and pH measurements are normalized
}
