% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_reads.R
\name{parse_reads}
\alias{parse_reads}
\title{Parse a read file}
\usage{
parse_reads(file, mthd = biotek_ten_wavelengths_universal)
}
\arguments{
\item{file}{Path to the input file}

\item{mthd}{the parsing method to be used}
}
\description{
This function parses a read file from a spectrophotometer
}
