% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHPredict.R
\name{pHPredict}
\alias{pHPredict}
\title{Predict the pH of wells in a microwell plate}
\usage{
pHPredict(
  read_file,
  model,
  wavelength_keyword = "wv",
  dye_name = "universal",
  ...
)
}
\arguments{
\item{read_file}{Path to the input file}

\item{model}{A function that takes parsed data as input and makes a prediction as an output}
}
\description{
This function parses a read file from a spectrophotometer and builds a model using a supplied training data set.
The function makes a prediction using the model using the read file data as an input.
}
