% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rDEI_table}
\alias{rDEI_table}
\title{Data frame with relative daily expression integral data}
\format{
A data frame with 24845 rows and 4 variables:
\describe{
  \item{ID}{TAIR gene identifier}
  \item{LDEQ}{relative daily expression integral of long day to equinox}
  \item{SDEQ}{relative daily expression integral of short day to equinox}
  \item{SDLD}{relative daily expression integral of short day to long day}
}
}
\usage{
rDEI_table
}
\description{
Data frame with relative daily expression integral data
}
\keyword{datasets}
