% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_makeFeature.R
\name{run_makeFeature}
\alias{run_makeFeature}
\title{Read gff3 files into Grange objects}
\usage{
run_makeFeature(gff3_paths, max_qvalue = 0.01, max_pvalue = 0.01, output_dir)
}
\arguments{
\item{gff3_paths}{Character vector. Full paths to gff3 files of motif mapping}

\item{max_qvalue}{Numeric.}

\item{max_pvalue}{Numeric.}

\item{output_dir}{Character. Output directory.}
}
\value{
Data frame.
}
\description{
Read gff3 files into Grange objects
}
