% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SO_lin_model}
\alias{SO_lin_model}
\title{Train a linear model with First order and Two way interaction and pure quadratic terms}
\usage{
SO_lin_model(abs, pH)
}
\arguments{
\item{abs}{a n x k data frame of absorbance measurements, where n is the number of observations, and k is the number of wavelengths}

\item{ph}{a normalized vector of length n of pH measurements for training}
}
\value{
a generated function that makes a prediction with the model with new data as an input
}
\description{
pH ~ rsm::SO(abs)
}
