% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train a model for pH prediction}
\usage{
train(abs_data, pH_vals, mdl_generator, wavelength_keyword = "wv", ...)
}
\arguments{
\item{abs_data}{a parsed absorbance dataframe}

\item{pH_vals}{pH values corresponding to each row of absorbance values}

\item{mdl_generator}{a generator function that produces a trained model and returns a prediction function, currently \code{gp_model} and \code{linear_model} are supported}
}
\description{
Use absorbance data and measured pH values to train a model.
}
